Drawing Envelope
(but it's wrong because of sample skiping)

+++ Source/Components/EnvelopeDrawer/Envelope.h
@@ -6,7 +6,7 @@ class envelope
 public:
 	struct envelope_node
 	{
-		envelope_node(int _sample, int _value): sample(_sample), value(_value)
+		envelope_node(int _sample, float _value): sample(_sample), value(_value)
 		{
 		}
